pyproject.toml
[tool.poetry]
name = "psy"
version = "0.1.0"
description = "Psychologist website with admin panel"
authors = ["Codex <codex@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.32"
alembic = "^1.13.2"
flask-wtf = "^1.2.1"
argon2-cffi = "^23.1.0"
boto3 = "^1.34.139"
gunicorn = "^22.0.0"
pillow = "^10.4.0"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
